feat(metadata): update article details and references
Update title, subtitle, author, keywords, subject, and category for
metadata.yaml to reflect the new article "Ввійти в IT". Adjusted
language settings to Ukrainian, updated fonts to DejaVu, and modified
reference section title to "Посилання".

Added a new online reference to bibliography.bib and enabled the
inclusion of uncited bibliography entries in the document.